Add proper API config for each agents
