+++ src/index.ts
@@ -7,9 +7,11 @@ import {
   setupSignalHandlers,
   startTaskExecutor,
   withApiLogger,
+  createApiServer,
 } from '@autonomys/agent-core';
 
 import { createAgentRunnerOptions } from './agent.js';
+import { createApiServerConfig } from './config/api.js';
 
 const configInstance = await getConfig();
 if (!configInstance) {
@@ -36,6 +38,8 @@ const orchestratorRunner = (() => {
 
 // Main function to run the application
 const main = async () => {
+  const apiServerConfig = createApiServerConfig(configInstance);
+  const _apiServer = createApiServer(apiServerConfig);
   try {
     const logger = createLogger('app');
     logger.info('Initializing orchestrator runner...');
